feat(bestiaire): le NC accueille le demi-niveau du livre (#218)
COF2 emploie ½ pour ses adversaires les plus faibles, et le dit jusque dans la
règle de l'attaque magique : « en ajoutant sa VOL à son NC (½ vaut 0) ». Bandit
de base, milicien, gobelin élite et rat géant sont des NC ½ dans le chapitre
Opposition. La colonne étant entière, ils étaient servis NC 1 — le double de
leur puissance annoncée, et donc deux fois trop cher dans le budget d'une
rencontre.

La colonne passe en flottant, des deux côtés du trait partagé : un MJ peut
désormais donner ½ à sa propre créature, comme le livre le fait pour les
siennes. Le formulaire accepte le demi-pas (`parseInt` tronquait 0,5 à 0) et
l'admin passe d'IntegerField à NumberField, qui aurait tronqué de même.

**½ se lit « ½ », pas « 0.5 ».** Le décimal est la forme de stockage ; `formatNC`
rend la notation du livre, sur la fiche, sur la carte et dans la recherche
globale.

**Le seuil du générateur suit.** `nc >= 1` excluait tout ce qui vaut moins d'un
niveau ; ces quatre créatures y entraient tant qu'elles étaient servies NC 1.
Leur rendre leur demi-niveau sans toucher au seuil les en aurait sorties sans
que personne ne le demande — d'où `NC_MINIMUM = 0.5`. Les NC 0 restent dehors :
ce sont des figurants, pas des adversaires.

**Un test intermittent réparé au passage.** Le contrôle des caractéristiques du
squelette visait la valeur par le voisinage dans le DOM (« le div après celui
qui dit INT ») et tombait une fois sur trois. Chaque case porte maintenant un
libellé qui réunit l'intitulé et la valeur — un lecteur d'écran annonçait « INT »
puis « ‑4 » sans dire qu'ils vont ensemble. Quatre exécutions de suite au vert.

Vérification : 8 tests unitaires (générateur avec ½ et sans NC 0, notation),
2 tests E2E, l'audit du bestiaire qui ne signale plus aucun écart. 576 tests
unitaires, 72 E2E, 106 tests backend, build, lint à zéro.

+++ backend/src/Entity/Trait/CreatureProfileTrait.php
@@ -25,9 +25,15 @@ trait CreatureProfileTrait
     #[Groups(['creature:read', 'creature:write', 'custom_creature:read', 'custom_creature:write'])]
     private ?string $description = null;
 
-    #[ORM\Column]
+    /**
+     * Niveau de créature. COF2 emploie ½ pour les adversaires les plus faibles (bandit,
+     * milicien, gobelin élite, rat géant) — « ½ vaut 0 » dans le calcul de l'attaque
+     * magique. Une colonne entière ne peut pas l'accueillir : ces créatures étaient servies
+     * NC 1, soit deux fois leur puissance annoncée dans le budget d'une rencontre.
+     */
+    #[ORM\Column(type: 'float')]
     #[Groups(['creature:read', 'creature:write', 'custom_creature:read', 'custom_creature:write'])]
-    private ?int $nc = null;
+    private ?float $nc = null;
 
     #[ORM\Column]
     #[Groups(['creature:read', 'creature:write', 'custom_creature:read', 'custom_creature:write'])]
@@ -83,8 +89,8 @@ trait CreatureProfileTrait
     public function getDescription(): ?string { return $this->description; }
     public function setDescription(?string $description): static { $this->description = $description; return $this; }
 
-    public function getNc(): ?int { return $this->nc; }
-    public function setNc(int $nc): static { $this->nc = $nc; return $this; }
+    public function getNc(): ?float { return $this->nc; }
+    public function setNc(float $nc): static { $this->nc = $nc; return $this; }
 
     public function getHp(): ?int { return $this->hp; }
     public function setHp(int $hp): static { $this->hp = $hp; return $this; }
